Gemfile: use "type-on-strap" as a Gem
There are two different ways to use the "type-on-strap" theme:

  - Forking it and making required changes.

  - Using it as a Ruby Gem.

We have picked the first option, which does provide the benefit that it
works well with GitHub Pages. But this way of doing things has several
issues:

  - It is hard to update the theme due to the mixture of upstream and
    downstream files.

  - It is hard to figure out which files are specific to your site.

  - The repository needlessly grows in size with objects that could have
    been provided by a Gem.

Jekyll nowadays provides a mechanism to set up a "remote_theme", which
is supposed to be compatible with GitHub Pages. So there aren't really
any reasons for why we should carry all the Gem-provided files, while
there are multiple reasons to not do so.

Convert the site to use "type-on-strap" as a Ruby Gem. Subsequent
commits will clean up files which are made obsolete by this change.