sass: reorganize includes for custom SCSS files
We have a couple of custom SCSS files that need to be included. This is
done by overriding the "_sass/type-on-strap.scss" file such that it
includes both the theme-provided files as well as our own ones. But this
requires us to manually keep the file in sync with any upstream changes
and is thus a bit tedious to use.

Refactor the layout such that we instead override "assets/css/main.css",
which only needs to include the "type-on-strap.scss" file as well as a
new "custom.scss" file that wires up remaining includes.