assets: reorganize informational letters
The informational letters are distributed across two different
directories, one for ".docx" and one for ".pdf" files, which clutters
the directory layout a bit. The only reason why that is the case is that
we use the directory path to filter out PDFs, only.

Refactor the code such that we directly filter for PDFs and merge both
filetypes into a single directory.