assets: provide proper responsive header feature images
The header feature image is of the wrong aspect ratio and thus exhibits
some issues depending on the current browser window's size. Furthermore,
we only provide a single size, which means that we always have to serve
the biggest variant regardless of the device type.

Provide properly-sized variants of the header image and enable the
responsive feature to fix this.