layouts/cards: make placeholders configurable per site
We're about to drop the tag mechanism in favor of using collections
exclusively. The "cards" layout uses the tags though to pick the default
placeholder image. This mechanism is weird even without the upcoming
rearchitecture as it requires the layout to know about its contents.

Drop the mechanism and instead make the default placeholder image
configurable per site.